+++ lib/src/responses.dart
@@ -1,3 +1,8 @@
+/// FromJsonDef<T>
+/// A helper method used to serialize the ContentChef content retrieved from the get or search api request
+///
+typedef FromJsonDef<T> = T Function(Map<String, dynamic> json);
+
 /// ContentChef responseMetadata class mapper
 ///
 class ResponseMetadata {
@@ -8,22 +13,42 @@ class ResponseMetadata {
   String publishedOn;
   List<String> tags;
 
-  /// Create a `ResponseMetadata` instance from a json object.
+  /// Create a `ResponseMetadata` instance
   ///
   /// Parameters:
-  /// - data: a json object to transform
-  /// - Returns: an instance of 'ResponseMetadata;
+  /// - authoringContentId: int (the content authoring unique id)
+  /// - contentLastModifiedDate: String (the content last modified date string)
+  /// - contentVersion: int (the published content version)
+  /// - id: int (the published content unique id)
+  /// - publishedOn: String (the content published date string)
+  /// - tags: List<String> (the content related tags)
+  /// - Returns: an instance of ResponseMetadata;
   ///
-  ResponseMetadata(Map<String, dynamic> data) {
-    authoringContentId = data['authoringContentId'];
-    contentLastModifiedDate = data['contentLastModifiedDate'];
-    contentVersion = data['contentVersion'];
-    id = data['id'];
-    publishedOn = data['publishedOn'];
-    tags = List<String>.from(data['tags']);
+  ResponseMetadata(
+      {this.authoringContentId,
+      this.contentLastModifiedDate,
+      this.contentVersion,
+      this.id,
+      this.publishedOn,
+      this.tags});
+
+  /// Create an instance of ResponseMetadata from a JSON object
+  ///
+  /// Parameters:
+  /// - json: the json object to serialize as a ResponseMetadata class
+  /// - Returns: an instance of ResponseMetadata
+  factory ResponseMetadata.fromJson(Map<String, dynamic> json) {
+    return ResponseMetadata(
+        authoringContentId: json['authoringContentId'] as int,
+        contentVersion: json['contentVersion'] as int,
+        id: json['id'] as int,
+        contentLastModifiedDate: json['contentLastModifiedDate'] as String,
+        publishedOn: json['publishedOn'] as String,
+        tags: List<String>.from(json['tags']));
   }
 
-  /// Method used to encode SortingField as JSON object
+  /// Method used to deserialize a ResponseMetadata class to JSON object
+  ///
   Map<String, dynamic> toJson() => {
         'authoringContentId': authoringContentId,
         'contentVersion': contentVersion,
@@ -34,7 +59,7 @@ class ResponseMetadata {
       };
 }
 
-/// ContentChef requestContents class mapper
+/// ContentChef RequestContents class mapper
 ///
 class RequestContext {
   String publishingChannel;
@@ -42,20 +67,38 @@ class RequestContext {
   String cloudName;
   String timestamp;
 
-  /// Create a `RequestContextInstance` instance from a json object.
+  /// Create a `RequestContextInstance` instance.
   ///
   /// Parameters:
-  /// - data: a json object to transform
+  /// - publishingChannel: String (request publishing channel)
+  /// - targetDate: String (request target date)
+  /// - cloudName: String (media cloud name reference)
+  /// - timestamp: String (request timestamp)
   /// - Returns: an instance of 'RequestContext'
   ///
-  RequestContext(Map<String, dynamic> data) {
-    publishingChannel = data['publishingChannel'];
-    targetDate = data['targetDate'];
-    cloudName = data['cloudName'];
-    timestamp = data['timestamp'];
+  RequestContext(
+      {this.publishingChannel,
+      this.targetDate,
+      this.cloudName,
+      this.timestamp});
+
+  /// Create an instance of RequestContext from a json Map
+  ///
+  /// Parameters:
+  /// - json: the json object to serialize as a RequestContext class
+  /// - Returns: an instance of RequestContext
+  ///
+  factory RequestContext.fromJson(Map<String, dynamic> json) {
+    return RequestContext(
+      publishingChannel: json['publishingChannel'] as String,
+      targetDate: json['targetDate'] as String,
+      cloudName: json['cloudName'] as String,
+      timestamp: json['timestamp'] as String,
+    );
   }
 
-  /// Method used to encode SortingField as JSON object
+  /// Method used to deserialize a RequestContext class to JSON object
+  ///
   Map<String, dynamic> toJson() => {
         'publishingChannel': publishingChannel,
         'targetDate': targetDate,
@@ -64,7 +107,7 @@ class RequestContext {
       };
 }
 
-/// ContentChef ContentResponse class mapper for a specified content T
+/// ContentChef ContentResponse class mapper for a specified content of type T
 ///
 class ContentResponse<T> {
   String definition;
@@ -76,24 +119,52 @@ class ContentResponse<T> {
   T payload;
   RequestContext requestContext;
 
-  /// Create a `ContentResponse<T>` instance from a json object.
+  /// Create a `ContentResponse<T>` instance.
   ///
-  /// Parameters:
-  /// - data: a json object to transform
+  ///  Parameters:
+  ///  - definition: String (content definition mnemonicId for the content)
+  ///  - repository: String (repository mnemonicId for the content)
+  ///  - publicId: String (content publicId)
+  ///  - offlineDate: String (content offline date)
+  ///  - onlineDate: String (content online date)
+  ///  - metadata: ResponseMetadata (instance of ResponseMetadata)
+  ///  - payload: T (instance of a class T)
+  ///  - requestContext: RequestContext (instance of RequestContext);
   /// - Returns: an instance of 'ContentResponse'
   ///
-  ContentResponse(Map<String, dynamic> data) {
-    definition = data['definition'];
-    repository = data['repository'];
-    publicId = data['publicId'];
-    metadata = ResponseMetadata(data['metadata']);
-    offlineDate = data['offlieDate'];
-    onlineDate = data['onlineDate'];
-    payload = data['payload'];
-    requestContext = RequestContext(data['requestContext']);
+  ContentResponse(
+      {this.definition,
+      this.repository,
+      this.publicId,
+      this.offlineDate,
+      this.onlineDate,
+      this.metadata,
+      this.payload,
+      this.requestContext});
+
+  /// Create an instance of ContentResponse from a json Map
+  ///
+  /// Parameters:
+  /// - json: the json object to serialize as a ContentResponse<T> class
+  /// - fromJson: FromJsonDef<T> function (used to serialize payload attribute as instance of T)
+  /// - Returns: an instance of RequestContext
+  ///
+  factory ContentResponse.fromJson(
+      Map<String, dynamic> json, FromJsonDef<T> fromJson) {
+    return ContentResponse(
+      definition: json['definition'] as String,
+      repository: json['repository'] as String,
+      publicId: json['publicId'] as String,
+      metadata: ResponseMetadata.fromJson(json['metadata']),
+      offlineDate: json['offlineDate'] as String,
+      onlineDate: json['onlineDate'] as String,
+      payload: fromJson(json['payload']),
+      requestContext: RequestContext.fromJson(json['requestContext']),
+    );
   }
 
-  /// Method used to encode SortingField as JSON object
+  /// Method used to deserialize a ContentResponse class to JSON object
+  ///
   Map<String, dynamic> toJson() => {
         'definition': definition,
         'repository': repository,
@@ -106,6 +177,8 @@ class ContentResponse<T> {
       };
 }
 
+/// ContentChef PaginatedResponse class mapper for a specified contents of type T
+///
 class PaginatedResponse<T> {
   List<ContentResponse<T>> items;
   int total;
@@ -113,19 +186,41 @@ class PaginatedResponse<T> {
   int take;
   RequestContext requestContext;
 
+  /// Create a `PaginatedResponse<T>` instance.
+  ///
+  ///  Parameters:
+  ///  - items: List<ContentResponse<T> (a list of ContentResponse<T> instance)
+  ///  - total: int (number of contents find in the search request)
+  ///  - skip: int (number of contents skipped in the search request)
+  ///  - take: int (number of contents taken in the search request)
+  ///  - requestContext: String (content online date)
+  ///  - Returns: an instance of 'PaginatedResponse<T>'
+  ///
+  PaginatedResponse({
+    this.items,
+    this.total,
+    this.skip,
+    this.take,
+    this.requestContext,
+  });
+
   /// Create a `PaginatedResponse<T>` instance from a json object.
   ///
   /// Parameters:
-  /// - data: a json object to transform
+  /// - json: the json object to serialize as a PaginatedResponse<T> class
+  /// - fromJson: FromJsonDef<T> function (used to serialize the items payload attribute as instance of T)
   /// - Returns: an instance of 'PaginatedResponse'
   ///
-  PaginatedResponse(Map<String, dynamic> data) {
-    items = List<ContentResponse<T>>.from(
-        data['items'].map((item) => ContentResponse(item)));
-    total = data['total'];
-    skip = data['skip'];
-    take = data['take'];
-    requestContext = RequestContext(data['requestContext']);
+  factory PaginatedResponse.fromJson(
+      Map<String, dynamic> json, FromJsonDef<T> fromJson) {
+    return PaginatedResponse(
+      items: List<ContentResponse<T>>.from(json['items']
+          .map((item) => ContentResponse<T>.fromJson(item, fromJson))),
+      total: json['total'] as int,
+      skip: json['skip'] as int,
+      take: json['take'] as int,
+      requestContext: RequestContext.fromJson(json['requestContext']),
+    );
   }
 
   /// Method used to encode SortingField as JSON object