+++ test/channels_test.dart
@@ -7,8 +7,24 @@ import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
 class MockedRequestExecutor extends Mock implements RequestExecutor {}
-class MockedContentResponse extends Mock implements ContentResponse {}
-class MockedPaginatedResponse extends Mock implements PaginatedResponse {}
+class MockedContentResponse extends Mock implements ContentResponse<MockedContentResult> {}
+class MockedPaginatedResponse extends Mock implements PaginatedResponse<MockedContentResult> {}
+class MockedContentResult {
+  String testAttribute;
+  String testAttribute2;
+
+  MockedContentResult({
+    this.testAttribute,
+    this.testAttribute2,
+  });
+
+  static MockedContentResult fromJson(Map<String, dynamic> json) {
+    return MockedContentResult(
+      testAttribute: json['testAttribute'] as String,
+      testAttribute2: json['testAttribute2'] as String,
+    );
+  }
+}
 
 void main () {
   final publicId = 'test-publicId';
@@ -20,7 +36,7 @@ void main () {
     final mockedGetResult = Future.value(MockedContentResponse());
     final mockedPaginatedResult = Future.value(MockedPaginatedResponse());
     group('OnlineChannel tests', () {
-      test('Initialize succesfully', () {
+      test('Initialize successfully', () {
         expect(OnlineChannel(
           config: config,
           apiKey: apiKey,
@@ -62,27 +78,51 @@ void main () {
         final getContentFilters = GetContentFilters(publicId: publicId);
         final requestPath = getOnlinePath(spaceId: config.spaceId, requestType: RequestTypes.content, channel: publishingChannel);
         final mockedRequestExecutor = MockedRequestExecutor();
-        when(mockedRequestExecutor.executeGetContentRequest(
+        when(mockedRequestExecutor.executeGetContentRequest<MockedContentResult>(
           path: requestPath,
           apiKey: apiKey,
           config: config,
-          filters: getContentFilters)
+          filters: getContentFilters,
+          fromJson: MockedContentResult.fromJson)
         ).thenAnswer((_) => mockedGetResult);
-        await OnlineChannel(config: config, apiKey: apiKey, publishingChannel: publishingChannel, requestExecutor: mockedRequestExecutor).getContent(filters: getContentFilters);
-        verify(await mockedRequestExecutor.executeGetContentRequest(path: requestPath, apiKey: apiKey, config: config, filters: getContentFilters)).called(1);
+        await OnlineChannel(
+          config: config,
+          apiKey: apiKey,
+          publishingChannel: publishingChannel,
+          requestExecutor: mockedRequestExecutor,
+        ).getContent<MockedContentResult>(filters: getContentFilters, fromJson: MockedContentResult.fromJson);
+        verify(await mockedRequestExecutor.executeGetContentRequest<MockedContentResult>(
+          path: requestPath,
+          apiKey: apiKey,
+          config: config,
+          filters: getContentFilters,
+          fromJson: MockedContentResult.fromJson
+        )).called(1);
       });
       test('searchContents method should call RequestExecutor.executeSearchContentsRequest method', () async {
         final searchContentsFilters = SearchContentsFilters(take: 0, skip: 0);
         final requestPath = getOnlinePath(spaceId: config.spaceId, requestType: RequestTypes.search, channel: publishingChannel);
         final mockedRequestExecutor = MockedRequestExecutor();
-        when(mockedRequestExecutor.executeSearchContentsRequest(
-            path: requestPath,
-            apiKey: publicId,
-            config: config,
-            filters: searchContentsFilters)
+        when(mockedRequestExecutor.executeSearchContentsRequest<MockedContentResult>(
+          path: requestPath,
+          apiKey: publicId,
+          config: config,
+          filters: searchContentsFilters,
+          fromJson: MockedContentResult.fromJson)
         ).thenAnswer((_) => mockedPaginatedResult);
-        await OnlineChannel(config: config, apiKey: apiKey, publishingChannel: publishingChannel, requestExecutor: mockedRequestExecutor).searchContents(filters: searchContentsFilters);
-        verify(await mockedRequestExecutor.executeSearchContentsRequest(path: requestPath, apiKey: apiKey, config: config, filters: searchContentsFilters)).called(1);
+        await OnlineChannel(
+          config: config,
+          apiKey: apiKey,
+          publishingChannel: publishingChannel,
+          requestExecutor: mockedRequestExecutor)
+        .searchContents<MockedContentResult>(filters: searchContentsFilters, fromJson: MockedContentResult.fromJson);
+        verify(await mockedRequestExecutor.executeSearchContentsRequest(
+          path: requestPath,
+          apiKey: apiKey,
+          config: config,
+          filters: searchContentsFilters,
+          fromJson: MockedContentResult.fromJson)
+        ).called(1);
       });
     });
 
@@ -90,7 +130,7 @@ void main () {
     group('PreviewChannel test', () {
       final stageStatus = PublishingStatus.stage;
 
-      test('Initilize succesfully', () {
+      test('Initialize successfully', () {
         expect(OnlineChannel(
           config: config,
           apiKey: apiKey,
@@ -137,27 +177,52 @@ void main () {
         final getContentFilters = GetContentFilters(publicId: publicId);
         final requestPath = getPreviewPath(spaceId: config.spaceId, requestType: RequestTypes.content, channel: publishingChannel, status: stageStatus);
         final mockedRequestExecutor = MockedRequestExecutor();
-        when(mockedRequestExecutor.executeGetContentRequest(
-            path: requestPath,
-            apiKey: apiKey,
-            config: config,
-            filters: getContentFilters)
+        when(mockedRequestExecutor.executeGetContentRequest<MockedContentResult>(
+          path: requestPath,
+          apiKey: apiKey,
+          config: config,
+          filters: getContentFilters,
+          fromJson: MockedContentResult.fromJson)
         ).thenAnswer((_) => mockedGetResult);
-        await PreviewChannel(status: stageStatus, config: config, apiKey: apiKey, publishingChannel: publishingChannel, requestExecutor: mockedRequestExecutor).getContent(filters: getContentFilters);
-        verify(await mockedRequestExecutor.executeGetContentRequest(path: requestPath, apiKey: apiKey, config: config, filters: getContentFilters)).called(1);
+        await PreviewChannel(
+          status: stageStatus,
+          config: config,
+          apiKey: apiKey,
+          publishingChannel: publishingChannel,
+          requestExecutor: mockedRequestExecutor,
+        ).getContent<MockedContentResult>(filters: getContentFilters, fromJson: MockedContentResult.fromJson);
+        verify(await mockedRequestExecutor.executeGetContentRequest<MockedContentResult>(
+          path: requestPath,
+          apiKey: apiKey,
+          config: config,
+          filters: getContentFilters,
+          fromJson: MockedContentResult.fromJson)
+        ).called(1);
       });
       test('searchContents method should call RequestExecutor.executeSearchContentsRequest method', () async {
         final searchContentsFilters = SearchContentsFilters(take: 0, skip: 0);
         final requestPath = getOnlinePath(spaceId: config.spaceId, requestType: RequestTypes.search, channel: publishingChannel);
         final mockedRequestExecutor = MockedRequestExecutor();
-        when(mockedRequestExecutor.executeSearchContentsRequest(
-            path: requestPath,
-            apiKey: apiKey,
-            config: config,
-            filters: searchContentsFilters)
+        when(mockedRequestExecutor.executeSearchContentsRequest<MockedContentResult>(
+          path: requestPath,
+          apiKey: apiKey,
+          config: config,
+          filters: searchContentsFilters,
+          fromJson: MockedContentResult.fromJson)
         ).thenAnswer((_) => mockedPaginatedResult);
-        await OnlineChannel(config: config, apiKey: apiKey, publishingChannel: publishingChannel, requestExecutor: mockedRequestExecutor).searchContents(filters: searchContentsFilters);
-        verify(await mockedRequestExecutor.executeSearchContentsRequest(path: requestPath, apiKey: apiKey, config: config, filters: searchContentsFilters)).called(1);
+        await OnlineChannel(
+          config: config,
+          apiKey: apiKey,
+          publishingChannel: publishingChannel,
+          requestExecutor: mockedRequestExecutor)
+        .searchContents<MockedContentResult>(filters: searchContentsFilters, fromJson: MockedContentResult.fromJson);
+        verify(await mockedRequestExecutor.executeSearchContentsRequest(
+          path: requestPath,
+          apiKey: apiKey,
+          config: config,
+          filters: searchContentsFilters,
+          fromJson: MockedContentResult.fromJson)
+        ).called(1);
       });
     });
   });