+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- add `fromJson` param in `GetContent` and `SearchContents` methods to help the user to serialize contents payload 
+
 ## 1.0.2
 
 - Change usage Uri.https in favour of Uri to better handle search query parameters as Iterable<String>